legalease-backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
motor==3.4.0
pymongo==4.7.3
python-jose[cryptography]==3.3.0
bcrypt==4.1.3
python-multipart==0.0.9
pymupdf==1.24.5
cloudinary==1.40.0
google-generativeai==0.7.2
python-dotenv==1.0.1
pydantic[email]==2.7.4
